+++ CHANGELOG.md
@@ -26,8 +26,13 @@ into pre-release sections below.
         * Use appropriate log-file path in collection-status endpoint (\#1902).
         * Drop `TaskUpdateV2.version` (\#1905).
         * Fix non-duplication check in `PATCH /api/v2/task-group/{id}/` (\#1911).
+        * Drop `TaskCreateV2.source` (\#1909).
+    * Dataset:
+        * Remove `TaskDumpV2.owner` attribute (\#1909).
     * Workflows and workflow tasks:
         * Drop `WorkflowTaskCreateV2.order` (\#1906).
+    * Admin:
+         * Remove owner from `GET admin/v2/task/` (\#1909).
 
 
 # 2.7.0a4 (unreleased)